Split the details view into attachment, copy, menu, action, and icon mixins.
